Add gameboard class and ability to add ships onto the gameboard.
In placeShip(), the method user can prescribe a length, row, column, and whether or not they want a vertical orientation for their ship. The gameboard will identify whether or not there is sufficient space for the ship, and then place it on there. Otherwise, null is returned. (Return ship or return null style follows in similar regard to the browser Element API)