Fix enumeration of player ships on roster and array synchronization
The player ships was not synced with the array correctly, searching up the wrong id. This is because the ships are not correctly ordered in the array. We fix this in #playerRegisterShips by changing the logic for registering ships. We use a for loop instead of forEach to ensure the ships are always ordered correctly.